fix: 持续修复测试编译错误，取得显著进展
重要修复:
- 修复 Decimal::new 参数类型转换问题 (u64 -> i64)
- 修复计费折扣结果类型错误 (Result<Decimal> vs Decimal)
- 修复 Property 测试中的 Strategy API 用法
- 添加缺失的类型导入: MetricType, InvoiceStatus
- 添加 gRPC 流类型的明确导入: TerminalStreamResponse 等
- 修复 chrono::Timelike trait 导入问题
- 修复字符串类型不匹配 (&str -> String)

进展统计:
✅ 测试编译错误: 259 → 227 → 204 → 198 (减少61个错误)
✅ 核心库编译: 完全成功 (0 错误)
⚡ 整体改进: 错误率下降 23.6%

下一步: 继续清理剩余198个测试编译错误

+++ tests/billing_comprehensive_tests.rs
@@ -345,7 +345,7 @@ mod data_integrity_tests {
         };
         
         // Average should be total / count
-        let expected_average = aggregated.total / Decimal::new(aggregated.count, 0);
+        let expected_average = aggregated.total / Decimal::new(aggregated.count as i64, 0);
         assert_eq!(aggregated.average, expected_average);
         
         // Maximum should be >= average >= minimum
